+++ dayz2/SpriteBatch.h
@@ -13,7 +13,7 @@ public:
 	~CSpriteBatch();
 	void begin(GLuint texture);
 	void end();
-	void draw(float x, float y, int width, int height, float u1, float v1, float u2, float v2, float rotation, glm::vec2 rotPoint);
+	void draw(float x, float y, float width, float height, float u1, float v1, float u2, float v2, float rotation, glm::vec2 rotPoint);
 	void drawVertices(float x1, float y1, float x2, float y2, float u1, float v1, float u2, float v2);
 	void draw(CSprite* pSprite);
 	void setColor(float r, float g, float b, float a);
